refactor: move dependency adding function into loader
