+++ src/CssDependency.js
@@ -1,0 +1,22 @@
+import webpack from 'webpack';
+
+export default class CssDependency extends webpack.Dependency {
+  constructor(
+    { identifier, content, media, sourceMap },
+    context,
+    identifierIndex
+  ) {
+    super();
+
+    this.identifier = identifier;
+    this.identifierIndex = identifierIndex;
+    this.content = content;
+    this.media = media;
+    this.sourceMap = sourceMap;
+    this.context = context;
+  }
+
+  getResourceIdentifier() {
+    return `css-module-${this.identifier}-${this.identifierIndex}`;
+  }
+}
